reformat rdrdt to use Rogers and Yau notation
