+++ PySDM/physics/drop_growth/howell_1949.py
@@ -2,6 +2,9 @@
 single-equation approximation of the vapour and heat diffusion problem
 as proposed in [Howell 1949](https://doi.org/10.1175/1520-0469(1949)006%3C0134:TGOCDI%3E2.0.CO;2)
 same as in [Mason 1951](https://doi.org/10.1088/0370-1301/64/9/307)
+
+The notation for terms associated with heat conduction and diffusion are from eq. 7.17
+ in [Rogers & Yau 1971](https://archive.org/details/shortcourseinclo0000roge_m3k2).
 """
 
 
@@ -9,11 +12,25 @@ class Howell1949:  # pylint: disable=too-few-public-methods
     def __init__(self, _):
         pass
 
-    # pylint: disable=too-many-arguments
     @staticmethod
-    def r_dr_dt(const, RH_eq, T, RH, lv, pvs, D, K):
-        return (
-            (RH - RH_eq)
-            / const.rho_w
-            / (const.Rv * T / D / pvs + lv**2 / K / T**2 / const.Rv)
-        )
+    def Fk(const, T, K, lv):
+        """thermodynamic term associated with heat conduction"""
+        return const.rho_w * lv / T / K * (lv / T / const.Rv)
+
+    @staticmethod
+    def Fd(const, T, D, pvs):
+        """the term associated with vapour diffusion"""
+        return const.rho_w * const.Rv * T / D / pvs
+
+    @staticmethod
+    def r_dr_dt(RH_eq, RH, Fk, Fd):
+        """Drop growth equation with radius r.
+
+        Parameters
+        ----------
+        Fk
+            Thermodynamic term associated with heat conduction from Rogers & Yau 1989.
+        Fd
+            Term associated with vapour diffusion.
+        """
+        return (RH - RH_eq) / (Fk + Fd)
