+++ docs/bibliography.json
@@ -740,7 +740,8 @@
     "usages": [
       "PySDM/physics/constants_defaults.py",
       "PySDM/physics/terminal_velocity/rogers_yau.py",
-      "PySDM/physics/drop_growth/rogers_yau_1989.py"
+      "PySDM/physics/drop_growth/howell_1949.py",
+      "PySDM/physics/drop_growth/mason_1971.py"
     ],
     "title": "A short course in clouds physics",
     "label": "Rogers & Yau 1989 (Pergamon Press)"
